+++ vulture/config.py
@@ -109,9 +109,10 @@ def _parse_args(args=None):
         metavar="PATTERNS",
         type=csv,
         default=missing,
-        help=f"Comma-separated list of paths to ignore (e.g.,"
-        f' "*settings.py,docs/*.py"). {glob_help} A PATTERN without glob'
-        f" wildcards is treated as *PATTERN*.",
+        help=f"Comma-separated list of path patterns to ignore (e.g.,"
+        f' "*settings.py,docs,*/test_*.py,venv"). {glob_help} A PATTERN without '
+        f" glob wildcards is treated as *PATTERN*. Patterns are matched"
+        f" against absolute paths.",
     )
     parser.add_argument(
         "--ignore-decorators",