+++ vulture/utils.py
@@ -67,9 +67,9 @@ def get_decorator_name(decorator):
 def get_modules(paths):
     """Retrieve Python files to check.
 
-    Loop over all given paths, abort if any ends with .pyc and add collect
-    the other given files (even those not ending with .py) and all .py
-    files under the given directories.
+    Loop over all given paths, abort if any ends with .pyc, add the other given
+    files (even those not ending with .py) and collect all .py files under the
+    given directories.
 
     """
     modules = []